Split IR.scala
