+++ core/src/main/scala-3/ir/Dialect.scala
@@ -1,0 +1,17 @@
+package scair.ir
+
+// ██╗ ██████╗░
+// ██║ ██╔══██╗
+// ██║ ██████╔╝
+// ██║ ██╔══██╗
+// ██║ ██║░░██║
+// ╚═╝ ╚═╝░░╚═╝
+
+/*≡==--==≡≡≡≡==--=≡≡*\
+||     DIALECTS     ||
+\*≡==---==≡≡==---==≡*/
+
+final case class Dialect(
+    val operations: Seq[OperationObject],
+    val attributes: Seq[AttributeObject]
+) {}
