+++ core/src/main/scala-3/ir/IRUtils.scala
