+++ core/src/main/scala-3/ir/Traits.scala
@@ -1,6 +1,4 @@
-package scair.traits
-
-import scair.ir.OpTrait
+package scair.ir
 
 // ████████╗ ██████╗░ ░█████╗░ ██╗ ████████╗ ░██████╗
 // ╚══██╔══╝ ██╔══██╗ ██╔══██╗ ██║ ╚══██╔══╝ ██╔════╝
@@ -8,6 +6,15 @@ import scair.ir.OpTrait
 // ░░░██║░░░ ██╔══██╗ ██╔══██║ ██║ ░░░██║░░░ ░╚═══██╗
 // ░░░██║░░░ ██║░░██║ ██║░░██║ ██║ ░░░██║░░░ ██████╔╝
 // ░░░╚═╝░░░ ╚═╝░░╚═╝ ╚═╝░░╚═╝ ╚═╝ ░░░╚═╝░░░ ╚═════╝░
+
+/*≡==--=≡≡≡=--=≡≡*\
+||    OPTRAIT    ||
+\*≡==---=≡=---==≡*/
+
+abstract class OpTrait {
+  def op: Operation
+  def trait_verify(): Unit = ()
+}
 
 /*≡==--=≡≡≡≡=--=≡≡*\
 ||   TERMINATOR   ||
